chore: add utility updateV2ClientTypeRef
